requirements.txt
telethon==1.32.1
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
APScheduler==3.10.4
Flask==3.0.0
Werkzeug==3.0.1
schedule==1.2.1
itchat-uos==1.5.0.dev0
pillow==10.2.0
